+++ Paragin.ExamAnalyzer.Cli/Students/Student.cs
@@ -1,9 +1,9 @@
 namespace Paragin.ExamAnalyzer.Cli.Student;
 
-internal sealed class Student(int number, string id, IReadOnlyList<decimal> questionScores)
+internal sealed class Student(int number, string id, IReadOnlyList<double> questionScores)
 {
     public int Number { get; } = number;
     public string Id { get; } = id;
-    public IReadOnlyList<decimal> QuestionScores { get; } = questionScores;
-    public decimal TotalScore { get; } = questionScores.Sum();
+    public IReadOnlyList<double> QuestionScores { get; } = questionScores;
+    public double TotalScore { get; } = questionScores.Sum();
 }
